+++ ScheduleCreator/controllers/CoursesController.java
@@ -34,6 +34,8 @@ import javafx.geometry.HPos;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.geometry.VPos;
+import javafx.scene.Node;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
@@ -639,6 +641,17 @@ public class CoursesController implements Initializable {
             Logger logger = Logger.getLogger(getClass().getName());
             logger.log(Level.SEVERE, "Failed to create new Window.", e);
         }
+    }
+    
+        public void changeToRegistrationScreen(ActionEvent _event) throws Exception {
+
+        Parent root = FXMLLoader.load(getClass().getResource("/ScheduleCreator/resources/views/registration_screen.fxml"));
+        Scene scene = new Scene(root);
+
+        //Get window object and refresh to show the new scene
+        Stage stage = (Stage) ((Node) _event.getSource()).getScene().getWindow();
+        stage.setScene(scene);
+        stage.show();
     }
 
 }
