+++ ScheduleCreator/resources/views/email_popup.fxml
@@ -10,9 +10,9 @@
 
 <VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" spacing="10.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ScheduleCreator.controllers.PopupController">
    <children>
-      <HBox alignment="CENTER" prefHeight="18.0" prefWidth="95.0" style="-fx-background-color: #66ccff;">
+      <HBox alignment="CENTER" prefHeight="18.0" prefWidth="95.0" style="-fx-background-color: #C0C0C0;">
          <children>
-            <Label text="Enter Your Email">
+            <Label text="Enter Your Email" textFill="#393939">
                <font>
                   <Font size="14.0" />
                </font>
