--- ScheduleCreator/resources/views/primary.fxml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane id="AnchorPane" prefHeight="560" prefWidth="1090" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ScheduleCreator.controllers.PrimaryController">
-    <children>
-         <Button layoutX="366.0" layoutY="256.0" onAction="#changeToSelectClasses" prefHeight="39.0" prefWidth="286.0" text="Schedule Generator">
-         <font>
-            <Font size="18.0" />
-         </font></Button>
-      <Button layoutX="366.0" layoutY="322.0" onAction="#changeToRegistrationScreen" prefHeight="39.0" prefWidth="286.0" text="My Registration Details">
-         <font>
-            <Font size="18.0" />
-         </font></Button>
-      <Text layoutX="349.0" layoutY="222.0" strokeType="OUTSIDE" strokeWidth="0.0" text="UNCG Schedule Creator" wrappingWidth="451.1748046875">
-         <font>
-            <Font name="System Italic" size="37.0" />
-         </font>
-      </Text>
-      <ImageView fitHeight="181.0" fitWidth="113.0" layoutX="696.0" layoutY="252.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../uncg_emblem_3-color-174x300.png" />
-         </image>
-      </ImageView>
-    </children>
-</AnchorPane>
